add auth check, change logging to log the user that called the method

+++ Milestone 3 Artifacts/Implementation/WebCrawlers.EdenFresh.UserManagement/UMManager.cs
@@ -11,6 +11,7 @@ namespace WebCrawlers.EdenFresh.UserManagement
         private Authorization authorization;
         ILogGateway dao;
         LogWriter writer;
+        Random rnd;
 
         public UMManager(int userID, UMService uMService, Authorization auth, LogWriter logging)
         {
@@ -19,40 +20,51 @@ namespace WebCrawlers.EdenFresh.UserManagement
             this.userId = userID;
             this.authorization = auth;
             this.writer = logging;
+            this.rnd = new Random();
         }
 
         public Boolean CreateAccount(String email, String password, Boolean isEnabled)
         {
+            if (!this.authorization.AuthorizeUser(this.userId))
+            {
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Business, $"User {this.userId} Is Unauthorized to Create Account");
+                return false;
+            }
             try
             {
                 stopwatch.Restart();
                 stopwatch.Start();
-                bool operationSuccessful = umService.CreateAccount(userId, email, password, isEnabled);
+                bool operationSuccessful = umService.CreateAccount(this.rnd.Next(100000, 999999), email, password, isEnabled);
                 stopwatch.Stop();
-                writer.Write(userId, LogWriter.LogLevel.Info, LogWriter.Category.DataStore, "Creating New User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Info, LogWriter.Category.DataStore, "Creating New User Account");
                 return (operationSuccessful && stopwatch.ElapsedMilliseconds <= 5000);
             }
             catch (Exception ex)
             {
-                writer.Write(userId, LogWriter.LogLevel.Error, LogWriter.Category.DataStore, "Error In Creating New User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.DataStore, "Error In Creating New User Account");
                 Console.WriteLine(ex.ToString());
                 return false;
             }
         }
         public Boolean DeleteAccount(int userId, String email, String password, Boolean isEnabled)
         {
+            if (!this.authorization.AuthorizeUser(this.userId))
+            {
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Business, $"User {this.userId} Is Unauthorized to Delete Account");
+                return false;
+            }
             try
             {
                 stopwatch.Restart();
                 stopwatch.Start();
                 bool operationSuccessful = umService.DeleteAccount(userId, email, password, isEnabled);
                 stopwatch.Stop();
-                writer.Write(userId, LogWriter.LogLevel.Info, LogWriter.Category.DataStore, "Deleting User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Info, LogWriter.Category.DataStore, "Deleting User Account");
                 return (operationSuccessful && stopwatch.ElapsedMilliseconds <= 5000);
             }
             catch(Exception ex)
             {
-                writer.Write(userId, LogWriter.LogLevel.Error, LogWriter.Category.DataStore, "Error In Deleting New User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.DataStore, "Error In Deleting New User Account");
                 Console.WriteLine(ex.ToString());
                 return false;
             }
@@ -60,18 +72,23 @@ namespace WebCrawlers.EdenFresh.UserManagement
 
         public Boolean UpdateAccount(int userId, String email, String password, Boolean isEnabled)
         {
+            if (!this.authorization.AuthorizeUser(this.userId))
+            {
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Business, $"User {this.userId} Is Unauthorized to Update Account");
+                return false;
+            }
             try
             {
                 stopwatch.Restart();
                 stopwatch.Start();
                 bool operationSuccessful = umService.UpdateAccount(userId, email, password, isEnabled);
                 stopwatch.Stop();
-                writer.Write(userId, LogWriter.LogLevel.Info, LogWriter.Category.Data, "Updating User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Info, LogWriter.Category.Data, "Updating User Account");
                 return (operationSuccessful && stopwatch.ElapsedMilliseconds <= 5000);
             }
             catch (Exception ex)
             {
-                writer.Write(userId, LogWriter.LogLevel.Error, LogWriter.Category.Data, "Error In Updating User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Data, "Error In Updating User Account");
                 Console.WriteLine(ex.ToString());
                 return false;
             }
@@ -79,6 +96,11 @@ namespace WebCrawlers.EdenFresh.UserManagement
 
         public Boolean EnableAccount(int userId, String email, String password)
         {
+            if (!this.authorization.AuthorizeUser(this.userId))
+            {
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Business, $"User {this.userId} Is Unauthorized to Enable Account");
+                return false;
+            }
             try
             {
                 stopwatch.Restart();
@@ -90,7 +112,7 @@ namespace WebCrawlers.EdenFresh.UserManagement
             }
             catch(Exception ex)
             {
-                writer.Write(userId, LogWriter.LogLevel.Error, LogWriter.Category.Data, "Error In Enabling User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Data, "Error In Enabling User Account");
                 Console.WriteLine(ex.ToString());
                 return false;
             }
@@ -98,6 +120,11 @@ namespace WebCrawlers.EdenFresh.UserManagement
 
         public Boolean DisableAccount(int userId, String email, String password)
         {
+            if (!this.authorization.AuthorizeUser(this.userId))
+            {
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Business, $"User {this.userId} Is Unauthorized to Disable Account");
+                return false;
+            }
             try
             {
                 stopwatch.Restart();
@@ -109,7 +136,7 @@ namespace WebCrawlers.EdenFresh.UserManagement
             }
             catch(Exception ex)
             {
-                writer.Write(userId, LogWriter.LogLevel.Error, LogWriter.Category.Data, "Error In Disabling User Account");
+                writer.Write(this.userId, LogWriter.LogLevel.Error, LogWriter.Category.Data, "Error In Disabling User Account");
                 Console.WriteLine(ex.ToString());
                 return false;
             }
